Add possibility to append distributionManagement configuration to pom.xml
Summary: We need to have one stop to handle 15-min related distributionManagement configuration for maven projects. Since it's (almost) not possible to do this via maven's settings.xml, we've decided to utilize the auto-release for this purpose, which will ensure the correct distribution configuration is set prior to publishing artifacts to jfrog.

Test Plan: `./gradlew clean test`

Reviewers: O5 elhub/devxp, michael.akinde

Reviewed By: O5 elhub/devxp, michael.akinde

Subscribers: michael.akinde, teamcity

JIRA Issues: TD-2537

Differential Revision: https://phabricator.elhub.cloud/D6034